商户bug修改

+++ src/main/java/com/xiliulou/electricity/vo/merchant/MerchantJoinRecordVO.java
@@ -63,5 +63,5 @@ public class MerchantJoinRecordVO {
     /**
      * 某个商户下的用户的数量
      */
-    private Long merchantUserNum;
+    private Integer merchantUserNum;
 }
